Work on a an interface for running a separate instance in an isolated process. This uses `DuplicateAndClose()` on Windows and `SCM_RIGHTS` on Linux.
This does not currently work, as neither Mono nor .Net Core has any methods to allow loading a file handle/socket descriptor to create a normal `Socket`.
The .Net Core simply does not support it (could perhaps be monkey hacked via SafeHandle and reflection), and Mono requires the handle to be registered via `_wsapi_handle_new_fd` which cannot be called from within the runtime.

An alternative would be to re-implement the `NetworkStream` object into something like `NativeSocketStream`, where we can read via `Syscal.recv`, but this requires additional logic to support TPL (async/await), essentially requiring us to use an even driven approach. If we go that way, maybe we should just use `libuv` as it already supports this.